% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelityBrierScore.R
\name{brierScore}
\alias{brierScore}
\title{Calculates Brier scores for a given matrix of predicted class
probabilities}
\usage{
brierScore(signatures, scale = TRUE)
}
\arguments{
\item{signatures}{data.frame containing the true classes and class prediction
probabilities for each signature. See details.}

\item{scale}{a boolean value: should the scores be scaled so 1 is the 
maximum and 0 corresponds to uniform predictions.}
}
\value{
a data.frame containing signature IDs, Brier scores
and sharpness scores
}
\description{
For a data.frame of signatures, class prediction probabilities and true 
classes, this function computes the Brier score [Hand, 1997]
for each signature and returns a data frame containing the results.
}
\details{
The data.frame \code{signatures} has the following columns: signatureID, 
truthClass, and multiple columns beginning with 'predictionClass' followed by
the class name. Each signature (denoted by 'signatureID') consists of many
observations, each of which has a true class membership (denoted by
'truthClass') and a corresponding probability value for membership in each 
class (denoted by 'predictionClass'+className). The rows of \code{signatures} 
correspond to each observation for which a classification is made.

Let \eqn{p_k, k = 1,\ldots,n}{p_k, k=1,...,n} indicate the class 
prediction probability vectors, where each \eqn{p_k}{p_k} corresponds 
to a single line in the \code{signatures}
data.frame and \eqn{n = nrow(signatures)}.  Let \eqn{J} denote the number of
classes. Let \eqn{\delta_{jk} = 1}{\delta_jk = 1} if 
and only if element \eqn{k} truly belongs to class \eqn{j}, 0 otherwise. Then
the Brier score is defined to be 
\deqn{B = \frac{1}{n}\sum_{k=1}^n\sum_{j=1}^J(\delta_{jk} - p_jk)^2}{B = (1/n)\sum_k \sum_j (\delta_jk - p_jk)^2}
The Brier score has a range of \eqn{[0,2]} where smaller scores are
better.

Strictly proper scoring rules can be decomposed into the sum of two 
components, one of which measures the sharpness of probability assignments
(i.e. the degree to which the probability assignment is placed on a single
or very few classes) and the other which is a measure of calibration or 
accuracy (i.e. the degree to which the probability assignments match the true
class).

We can define the component of the Brier score decomposition that measures 
the sharpness of probability predictions thusly:
\deqn{S = -(1/n)\sum_{k=1}^n\sum_{j=1}^J(p_jk)(1 - p_jk)}{S = -(1/n)\sum_k \sum_j (p_jk)(1-p_jk)}
The Brier score and sharpness are returned by this function.

To facilitate side-by-side comparison with the log score (see
\code{\link{logScore}}) we can transform the Brier score so that the maximum
value is 1 and a uniform probability assignment 
gives a score of 0. The unscaled Brier score for uniform probability 
assignments is \eqn{B = (J-1)/J}. Then the scaled Brier score is given by 
\deqn{B^R = 1+J/(1-J)*B}
The scaled Brier score takes values in \eqn{[1-J^2/(J-1), 1]}.
The sharpness scores are scaled in nearly the same manner, except that the 
coefficient is the negative of that in the previous equation:
\deqn{S^R = 1+J/(J-1)*S}
If \code{scale=TRUE} 
the scaled Brier score and scaled sharpness are returned.
}
\author{
Amanda White
}
\references{
Hand, D.J. (1997). Construction and Assessment of Classification 
Rules. John Wiley & Sons.
}

