% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exClassifed.R
\docType{data}
\name{exClassified}
\alias{exClassified}
\title{Example data (TODO:  Document this)}
\format{TODO}
\description{
Some more description
}
\examples{
data(exClassifed)
str(exClassified)
}
\keyword{datasets}

