% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelityDiscrete.R
\name{fidelityDiscrete}
\alias{fidelityDiscrete}
\title{Calculate the a discrete loss (or utility) for a categorical classifier}
\usage{
fidelityDiscrete(signatures, lossMatrix, truthClassLabel = "truthClass",
  predClassLabel = "predictedClass", lossLabel = "loss")
}
\arguments{
\item{signatures}{A data frame containing the predicted and true classes for each observation.
See details.}

\item{lossMatrix}{A matrix containing values of the loss (or utility) for each
combination of truth class and predicted class.  The predicted class labels should be the column names
and the truth class labels should be the rownames.  The matrix should not contain any missing values.}

\item{truthClassLabel}{A character string indicating the name of the column in \code{signatures} that
contains the truth labels.}

\item{predClassLabel}{A character string indicating the name of the column in \code{signatures} that contains the
predicted labels.}

\item{lossLabel}{A character string indicating the name of the column that will be added
to the \code{signatures} data frame containing the loss.}
}
\value{
The \code{signatures} data frame, with an additional column named \code{lossLabel} which has
the loss for each observation.
}
\description{
Calculate the a discrete loss (or utility) for a categorical classifier
}
\details{
In the \code{signatures} data frame, each row corresponds to an observation, or instance,
for which the loss will be calculated.  It must contain at least two columns, one for the
truth labels and one for the predicted labels.  The names of these columns are specifed by
\code{truthClassLabel} and \code{predictedClassLabel}.
}
\examples{

# A an example of 3 truth classes being mapped to one of the same 3 classes
# Some truth data
truthData <- sample(rep(letters[1:3], each = 4))

# Some predicted data
predData <- sample(rep(letters[1:3], each = 4))

# The loss matrix
lossM <- matrix(c(0,2,1,1,0,3,1,5,0), nrow = 3, ncol = 3,
               dimnames = list(letters[1:3], letters[1:3]))

lossM

# Calculate the loss
fidelityDiscrete(data.frame(truthClass = truthData, predictedClass = predData),
                 lossM)

# An assymetric example where 3 classes are mapped to 2
# An assymetric loss matrix, where predicted data are binary                   
lossM <- matrix(c(0,2,4,5,0,0), nrow = 3, ncol = 2,
                dimnames = list(letters[1:3], letters[4:5]))

lossM

# Binary predicted data
newData <- data.frame(truthClass = truthData, predictedClass = sample(rep(letters[4:5], each = 6)))

# Calculate the loss
fidelityDiscrete(newData, lossM)

# An example with missing values
newData1 <- newData
newData1[2, 1] <- NA
newData1[4, 2] <- NA
newData1[7, ] <- NA
newData1

fidelityDiscrete(newData1, lossM)
}
\author{
Landon Sego
}

