% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelityLogScore.R
\name{fidelityLogScore}
\alias{fidelityLogScore}
\title{Calculate the log score for each observation in a dataframe}
\usage{
fidelityLogScore(signatures, lowerBoundCoef = 0.5, base = exp(1),
  keepPreds = FALSE)
}
\arguments{
\item{signatures}{data.frame containing the true classes and class prediction
probabilities for each signature. See details.}

\item{lowerBoundCoef}{A positive fraction, in [0, 1], that is
multiplied by the smallest nonzero true-class probability, thereby defining
a positive lower bound of probabilities belonging to the true class and avoiding
log scores of \code{-Inf}.
This lower bound probability is substituted for cases where the probability of
the true class is 0.  If \code{lowerBoundCoef = 0} and if the probability of the true class
is 0, the resulting \code{-Inf} is left unchanged.}

\item{base}{The base of the logarithm. Defaults to the natural logarithm.}

\item{keepPreds}{A logical indicating whether the columns containing the probabilities
for each class (the \code{predictionClass} columns)
should be retained or not in the output dataframe.}
}
\value{
The \code{signatures} dataframe is returned with an additional column,
\code{logScore}, which is the logarithm of the probability of the true class. The
\code{predictionClass} columns are retained if \code{keepPreds = TRUE}.
}
\description{
Calculate the log score for each observation in a dataframe, probably as a
precursor to calculating the utility function for fidelity when classifier
output is a vector of probabilities.
}
\details{
The data.frame \code{signatures} has the following columns: signatureID, 
truthClass, and multiple columns beginning with 'predictionClass' followed by
the class name. Each signature (denoted by 'signatureID') consists of many
observations, each of which has a true class membership (denoted by
'truthClass') and a corresponding probability value for membership in each 
class (denoted by 'predictionClass'+className). The rows of \code{signatures} 
correspond to each observation for which a classification is made.

The log score is simply the logarithm of the probability that is assigned by
the classifier to the true class.
The log score ranges from \eqn{(-\infty,0]} where larger scores are
better. Typically we would expect the log sore to be transformed via a linear
transformation to make it a utility function.
}
\examples{
data(exampleScoreData)
y <- fidelityLogScore(exampleScoreData)
head(y)
}
\author{
Landon Sego
}

