% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBootstrapIndexes.R
\name{genBootstrapIndexes}
\alias{genBootstrapIndexes}
\title{Generate a set indexes which define a single bootstrap sample for each signature}
\usage{
genBootstrapIndexes(signatureID)
}
\arguments{
\item{signatureID}{A vector of signature labels}
}
\value{
An integer vector of indexes corresponding to signatureID that indicate
the elements of the bootstrap sample
}
\description{
Generate a set of indexes which define a single bootstrap sample in a stratified manner
where each signature ID is a strata,
and a bootstrap sample is generated separately for each signature ID. This function is
a helper to other functions in the SQM package, like \code{\link{calcScore}}.
}
\examples{
# A single sample
sigLabels <- rep(letters[1:3], each = 4)
print(sigLabels)
SQM:::genBootstrapIndexes(sigLabels)
}

