% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idParetoFrontier.R
\name{idParetoFrontier}
\alias{idParetoFrontier}
\title{Identifies the pareto frontier}
\usage{
idParetoFrontier(X, orientation, tol = 1e-10)
}
\arguments{
\item{X}{A data frame containing the attributes on the columns and one row
for each value of \code{signatureID}, where the required column
\code{signatureID} indicates the diferent signature systems under
comparison.}

\item{orientation}{A named vector of 1's and -1's, where the names identify the columns
in \code{X} that will be used in the pareto analysis. A value of \code{1} indicates the
utility of the variable is increasing, i.e. larger values of the variable are preferred.
A value of \code{-1} indicates the utility of the variable is decreasing, i.e., smaller values
of the variable are preferred to larger ones.}

\item{tol}{the tolerance for determining if two instances are tied.
If the absolute difference of a particular attribute exceeds \code{tol},
they are considered to be different (not tied).}
}
\value{
A data frame identical to \code{X}, except with a new column labeled \code{frontier} 
indicating if the particular \code{signatureID} is on the frontier.
}
\description{
Identifies the pareto frontier for a set of summarized signatures
}
\details{
Identify the pareto frontier by 
the attributes. A \code{signatureID} dominates another if it is better than the
other \code{signatureID} for at least one attribute,
and, at a minumum, equivalent for all the other attributes.

The new varible \code{frontier} is of type character, each element of which can
take one of four possible values:
}
\examples{
data(bacillusAttributes)
p <- idParetoFrontier(bacillusAttributes, c(risk = -1, ss = -1, cost = -1))
print(p)
plot(p)
}
\author{
Landon Sego
}

