% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Identify the pareto frontier of a set of attributes}
\usage{
pareto(testData, orientation, weights = NULL)
}
\arguments{
\item{testData}{A data frame containing attributes used to distinguish
the signature systems on the columns
and observations on the rows. If \code{testData} has column called \code{signatureID},
the expected utility is calculated separately for each \code{signatureID}.}

\item{orientation}{A named vector of 1's and -1's, where the names identify the columns
in \code{testData} that will be used in the pareto analysis. A value of \code{1} indicates the
utility of the variable is increasing, i.e. larger values of the variable are preferred.
A value of \code{-1} indicates the utility of the variable is decreasing, i.e., smaller values
of the variable are preferred to larger ones.}

\item{weights}{A numeric vector of non-negative values that approximate the joint
distribution of all the variables over which the expectation will be calculated.
\code{weights} should have the same number of elements as the number of rows in
\code{testData}.
Alternatively, this can be a character string indicating the name of a column in \code{testData} which
contains the weights. If \code{NULL}, equal weight is given to each observation.}
}
\value{
A data frame with the expected value of each attribute, calculated for
each \code{signatureID},
and an additional column called \code{frontier} indicating which \code{signatureID}s
constitute the pareto frontier.
}
\description{
Identify the pareto frontier of a set of attributes
}
\details{
To summarize across multiple observations within a \code{signatureID}, the expected value
of each attribute named in \code{orientation} is calculated using the \code{weights}.

Weights are normalized for each \code{signatureID}.

The \code{frontier} column is generated using \code{\link{idParetoFrontier}}.
}
\examples{

# Cook up some data
someData <- data.frame(signatureID = rep(LETTERS[1:5], each = 5),
                       a1 = runif(25, -2, 4),
                       a2 = rpois(25, 3), row.names = letters[1:25],
                       weights = rep(1:5, 5))

# Identify the pareto frontier
p <- pareto(someData, c(a1 = -1, a2 = 1), weights = "weights")
print(p)
plot(p)
}
\author{
Landon Sego
}

