% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saUtil.R
\name{saUtil}
\alias{plot.saUtilCall}
\alias{plot.saUtilFun}
\alias{print.saUtilCall}
\alias{print.saUtilFun}
\alias{saUtil}
\title{Generates or evaluates a single attribute utility function}
\usage{
saUtil(saUtilMethod = "expUtil", z = NULL, ...)

\method{print}{saUtilCall}(x, ...)

\method{print}{saUtilFun}(x, ...)

\method{plot}{saUtilFun}(x, ...)

\method{plot}{saUtilCall}(x, ...)
}
\arguments{
\item{saUtilMethod}{A character string (or unique abbreviation) indicating the single attribute
utility function method. Must be one (and only one) of the method functions returned by \code{\link{saUtilMethods}}.}

\item{z}{A numeric vector of attribute values.}

\item{x}{a \code{saUtilCall} or \code{saUtilFun} object returned by \code{saUtil}}

\item{dots}{Optional named arguments to the \code{\link{saUtilMethods}} functions that will set the parameters.}

\item{dots}{additional arguments passed to \code{\link{print.default}} or \code{\link{plot.default}}}
}
\value{
\item{1}{If \code{z = NULL}, the single attribute utility function with fixed parameters is returned
and can be called later. This function has a single argument, \code{z}, which are the measured
values of the attribute of interest. The returned object is of class \code{saUtilFun}.}
\item{2}{If a numeric vector is provided for the argument \code{z},
the utility method is evaluated using \code{z} for the given parameters \code{...} and the
calculated utility values are returned. The returned object is of class \code{saUtilCall}.}
}
\description{
Generate or evaluaties a single attribute utility function
}
\examples{
# Create a decreasing, exponential single attribture utility function
uf <- saUtil("exp", zrange = c(2, 14), urange = c(1, 0), certEquiv = 5)
is.function(uf)

# Info about 'uf' is contained in the attributes
attributes(uf)

# Plotting without specific data values
plot(uf)

# Now calculate the utility for a sequence of points
u <- uf(seq(2, 14, length = 10))
print(u)
plot(u)

# If we send actual data (attribute values) into saUtil(), the utility function
# is calculated for us
calculatedUtil <- saUtil("logUtil", z = runif(10), shift = -0.01, zrange = c(1/10, 1))
print(calculatedUtil)
plot(calculatedUtil)

# An example where the range of attribute values is defined by the existing data
calculatedUtil <- saUtil("logUtil", z = runif(10), shift = -0.01)
plot(calculatedUtil, col = "Blue", pch = 2, cex = 2)
}

