% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMethods.R
\name{scoreMethods}
\alias{scoreMethods}
\title{Lists the SQM classification score methods.}
\usage{
scoreMethods(method = NULL)
}
\arguments{
\item{method}{a character string for an score method. If no value is
specified, then all available score methods are returned.}
}
\value{
By default, a vector with all available score methods.
}
\description{
This function enumerates the score methods available in the SQM package.
}
\details{
By default, this function returns a vector with all of the available score
methods that are used within the SQM package. This function also acts as a
lookup to determine if a particular score method is available. If 'method'
is specified, then a lookup is performed, and either TRUE or FALSE is returned
to indicate the presence or absence of the given 'method, respectively.
}

