% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQM-package.R
\docType{package}
\name{SQM}
\alias{.onLoad}
\alias{SQM}
\alias{SQM-package}
\alias{package-SQM}
\title{The R Implementation of the SQM Analytic Framework}
\usage{
.onLoad(libname, pkgname)
}
\description{
The SQM package contains the R code for the Signature Quality Metrics project
under the Signature Discovery Initiative (SDI) at Pacific Northwest
National Laboratory (PNNL). The purpose of the SQM project is to provide
subject-matter experts (SMEs) with a set of tools to assess the quality of a
set of signatures in terms of accuracy, cost, risk, and utility. For each of
these components, we have provided easy-to-use, accessible functions to
measure the quality of signatures.
}

